+++ pkg/util/net.go
@@ -7,6 +7,7 @@ import (
 	"math"
 	"math/big"
 	"net"
+	"os"
 	"strconv"
 	"strings"
 	"time"
@@ -620,4 +621,21 @@ func UDPConnectivityListen(address string) error {
 			continue
 		}
 	}
+}
+
+func GetDefaultListenPort() string {
+	addr := "0.0.0.0"
+	if os.Getenv("ENABLE_BIND_LOCAL_IP") == "true" {
+		podIpsEnv := os.Getenv("POD_IPS")
+		podIps := strings.Split(podIpsEnv, ",")
+		// when pod in dual mode, golang can't support bind v4 and v6 address in the same time,
+		// so not support bind local ip when in dual mode
+		if len(podIps) == 1 {
+			addr = podIps[0]
+			if CheckProtocol(podIps[0]) == kubeovnv1.ProtocolIPv6 {
+				addr = fmt.Sprintf("[%s]", podIps[0])
+			}
+		}
+	}
+	return addr
 }
